pyproject.toml
[tool.poetry]
name = "ecommerce-mvp"
version = "0.1.0"
description = "MVP e-commerce data platform for repeat purchase prediction"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.0"
featuretools = "^1.28.0"
scikit-learn = "^1.3.0"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "0.34.0"}
joblib = "^1.3.0"
kafka-python = "^2.0.2"
faker = "^19.0.0"
feast = "^0.49.0"
redis = "^4.6.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.4.0"
jupyter = "^1.0.0"
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
